finalisation de toutes les étapes > début de la dernière étape (fonctionnalité de like)

+++ backend/controllers/sauce.js
@@ -1,5 +1,8 @@
 const Sauce = require('../models/sauce');
 
+const fs = require('fs');
+const sauce = require('../models/sauce');
+
 exports.getAllSauce = (req, res, next) => {
     Sauce.find()
         .then(sauces => res.status(200).json(sauces))
@@ -37,10 +40,10 @@ exports.modifySauce = (req, res, next) => {
     Sauce.findOne({_id: req.params.id})
         .then((sauce) => {
             if (sauce.userId != req.auth.userId) {
-                res.status(401).json({ message : 'Not authorized'});
+                res.status(401).json({ message : 'Non autorisé'});
             } else {
                 Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
-                .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                .then(() => res.status(200).json({message : 'Objet modifié !'}))
                 .catch(error => res.status(401).json({ error }));
             }
         })
@@ -50,7 +53,20 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id})
+       .then(sauce => {
+           if (sauce.userId != req.auth.userId) {
+               res.status(401).json({message: 'Non autorisé'});
+           } else {
+               const filename = sauce.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Sauce.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
